# flyio-admin/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "flyio-admin API",
    "description": "Application and Orchestration layer API for Flyio.ai — CMS, Blog, Trips, Knowledge Base, Web Scraper coordination, and LLM Workflow Orchestration.",
    "version": "1.0.0",
    "contact": {
      "name": "Flyio.ai Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT token acquired via POST /api/admin/login"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "admin" },
          "password": { "type": "string", "example": "admin1234" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6..." },
          "admin": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 1 },
              "username": { "type": "string", "example": "admin" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "summary": "Admin Login",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/admin/me": {
      "get": {
        "summary": "Get authenticated admin profile",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Admin profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/blog/posts": {
      "get": {
        "summary": "Public blog posts with sorting",
        "tags": ["Blog"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["created_at", "author"], "default": "created_at" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": {
          "200": { "description": "Paginated blog posts list" }
        }
      }
    },
    "/api/blog/posts/{slug}": {
      "get": {
        "summary": "Public single blog post by slug",
        "tags": ["Blog"],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Blog post detail" },
          "404": { "description": "Post not found" }
        }
      }
    },
    "/api/admin/blog/posts": {
      "get": {
        "summary": "List admin blog posts",
        "tags": ["Blog Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Posts list" } }
      },
      "post": {
        "summary": "Create new blog post",
        "tags": ["Blog Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Post created" } }
      }
    },
...
  }
}